fix various error in eole + fix warnings + fix pipsmake declarations
git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@15161 32a0c1e6-9308-0410-80dd-f9d2257a62c7

+++ src/Libs/expressions/instruction_selection.c
@@ -55,7 +55,7 @@ inst_sel_ctx;
  * if not, NIL is returned.
  */
 static list /* of expression */ 
-is_this_op(expression e, entity op, int len)
+is_this_op(expression e, entity op, size_t len)
 {
 	if (expression_call_p(e))
 	{
